refactor(common): migrate core currency consumers
